Move MessageGuard out to storage/common sub-library
