+++ storage/src/vespa/storage/distributor/top_level_bucket_db_updater.h
@@ -4,11 +4,11 @@
 #include "bucketlistmerger.h"
 #include "distributor_component.h"
 #include "distributormessagesender.h"
-#include "messageguard.h"
 #include "operation_routing_snapshot.h"
 #include "outdated_nodes_map.h"
 #include "pendingclusterstate.h"
 #include <vespa/document/bucket/bucket.h>
+#include <vespa/storage/common/message_guard.h>
 #include <vespa/storage/common/storagelink.h>
 #include <vespa/storageapi/message/bucket.h>
 #include <vespa/storageapi/messageapi/messagehandler.h>
