+++ storage/src/vespa/storage/storageserver/mergethrottler.cpp
@@ -1,6 +1,7 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "mergethrottler.h"
+#include <vespa/storageapi/message/state.h>
 #include <vespa/storage/common/nodestateupdater.h>
 #include <vespa/storage/common/dummy_mbus_messages.h>
 #include <vespa/storage/persistence/messages.h>
