+++ storage/src/vespa/storage/storageserver/mergethrottler.h
@@ -8,9 +8,9 @@
 #pragma once
 
 #include <vespa/storage/config/config-stor-server.h>
+#include <vespa/storage/common/message_guard.h>
 #include <vespa/storage/common/storagelink.h>
 #include <vespa/storage/common/storagecomponent.h>
-#include <vespa/storage/distributor/messageguard.h>
 #include <vespa/storageframework/generic/status/htmlstatusreporter.h>
 #include <vespa/storageapi/message/bucket.h>
 #include <vespa/document/bucket/bucket.h>
